+++ src/game/model/GameModel.java
@@ -1,8 +1,5 @@
 package game.model;
 
-import java.io.IOException;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
 import java.io.Serializable;
 
 /**
