+++ src/game/model/skill/Dot.java
@@ -1,5 +1,10 @@
 package game.model.skill;
 
+import java.util.Timer;
+import java.util.TimerTask;
+
+import game.model.Map;
+import game.utilities.Vector2D;
 import game.utilities.ViewSettings;
 
 /**
@@ -9,7 +14,7 @@ import game.utilities.ViewSettings;
  * @see {@link Skill}
  *
  */
-public abstract class Dot extends Skill {
+public class Dot extends Skill {
 	
 	//****************************** Attributes ******************************
 	
@@ -48,5 +53,53 @@ public abstract class Dot extends Skill {
 	protected int getCount() {
 		return count;
 	}
+
 	
+	@Override
+	protected boolean usable(SkillUser user) {
+			setTarget(null);
+			Map map = user.getCurrentMap();
+			Vector2D direction = user.getDirectionFacing();
+			Vector2D targetPos = new Vector2D(user.getX(), user.getY());
+			int a = 1;
+			while(a <= getRange()) {
+				targetPos.add(direction);
+				if(map.isTargetAt(targetPos.getIntX(), targetPos.getIntY())) {
+					setTarget(map.getTargetAt(targetPos.getIntX(), targetPos.getIntY()));
+					break;
+				}
+				else if(!map.noCollision(targetPos.getIntX(), targetPos.getIntY()))
+					break;
+				a++;
+			}
+		return (getTarget() != null);
+	}
+	
+	@Override
+	public void use(SkillUser user) {
+		Skill skillToNotify = this;
+		if (timingIsOk()){
+			if(usable(user)) {
+			Timer timer = new Timer();
+			long startTime=System.currentTimeMillis();
+			setLastExecutionTime (startTime);
+			timer.scheduleAtFixedRate(new TimerTask() {
+				  @Override
+				  public void run() {
+						setPosition(getTarget().getX(), getTarget().getY());
+						user.notifyObservers(skillToNotify);
+						getTarget().loseHp(getDamage());
+						if(getTarget().isDead()){
+							gainKillXp(user);
+							timer.cancel();
+					        timer.purge();
+					   	}
+						else if(System.currentTimeMillis() - startTime > getCount()*getLapse()){
+							timer.cancel();
+					        timer.purge();
+						}
+				  }
+			}, 0, getLapse());}
+		}
+	}
 }
